+++ R/flextable.R
@@ -2,29 +2,30 @@
 
 #' Turns a `crosstable` object into a formatted `flextable`
 #'
-#' @param x the result of [crosstable()]
-#' @param keep_id whether to keep the `.id` column
-#' @param by_header a string to override the `by` header
-#' @param autofit whether to use [flextable::autofit()] on the table
-#' @param compact whether to compact the table
-#' @param show_test_name in the `test` column, show the test name
-#' @param fontsizes font sizes as a list of keys `c(body, subheaders, header)`. If set through arguments instead of options, all 3 names should be specified.
-#' @param padding_v vertical padding (body)
-#' @param remove_header_keys if `TRUE` and `x` has several `by` strata, header will only display values
-#' @param header_show_n show the size (`N=xx`) of each terminal `by` strata in their header
+#' @param x the result of [crosstable()].
+#' @param keep_id whether to keep the `.id` column. 
+#' @param by_header a string to override the header if `x` has only one `by` stratum.
+#' @param autofit whether to use [flextable::autofit()] on the table.
+#' @param compact whether to compact the table. If `TRUE`, see [compact.crosstable()] to see how to use `keep_id`.
+#' @param show_test_name in the `test` column, show the test name.
+#' @param fontsizes font sizes as a list of keys. Default to `list(body=11, subheaders=11, header=11)`. If set through arguments instead of options, all 3 names should be specified.
+#' @param padding_v vertical padding (body).
+#' @param remove_header_keys if `TRUE` and `x` has several `by` strata, header will only display values.
+#' @param header_show_n numeric vector telling on which depth the group size should be indicated in the header. You can control the pattern using option `crosstable_options`. See [crosstable_options()] for details about it. See example for use case.
+#' @param header_show_n_pattern glue pattern used when `header_show_n==TRUE`. `.col` is the name of the column and `.n` the size of the group. Default to `{.col} (N={.n})`; you can use `{.col_key}` and `{.col_val}` when `by` has multiple stratum.
 #' @param generic_labels names of the crosstable default columns. Useful for translation for instance. 
-#' @param ... unused
+#' @param ... unused.
 #' 
-#' @return a flextable
+#' @return a flextable.
 #'
 #' @author Dan Chaltiel
 #' @aliases ctf cross_to_flextable to_flextable
 #' @describeIn as_flextable Turns a `crosstable` object into a formatted `flextable`.
 #' @seealso [crosstable()], [flextable::flextable()], [as_gt.crosstable()]
 #' 
-#' @importFrom dplyr select lead sym %>%
+#' @importFrom dplyr %>% select lead sym recode
 #' @importFrom stringr str_replace str_replace_all str_remove
-#' @importFrom flextable flextable autofit add_header_row merge_v merge_h bold align hline_top hline_bottom border_inner_h hline fix_border_issues padding as_flextable fontsize vline vline_left vline_right set_header_df
+#' @importFrom flextable flextable autofit add_header_row set_header_labels merge_v merge_h bold align hline_top hline_bottom border_inner_h hline fix_border_issues padding as_flextable fontsize vline vline_left vline_right set_header_df
 #' @importFrom officer fp_border
 #' @importFrom checkmate assert_class vname
 #' @importFrom tibble as_tibble
@@ -39,14 +40,14 @@
 #'                    crosstable_fontsize_subheaders=10, 
 #'                    crosstable_fontsize_body=8)
 #' crosstable(iris) %>% as_flextable()
-#' crosstable(mtcars2, by=c(am, vs)) %>% as_flextable()
+#' crosstable(mtcars2, -model, by=c(am, vs)) %>% as_flextable(header_show_n=1:2)
 #' crosstable(mtcars2, cols=c(mpg, cyl), by=am, effect=TRUE) %>% 
 #'    as_flextable(keep_id=TRUE, autofit=FALSE)
 #' crosstable(mtcars2, cols=c(mpg, cyl), by=am, effect=TRUE) %>% 
-#'    as_flextable(compact=TRUE)
+#'    as_flextable(compact=TRUE, header_show_n=TRUE)
 #' 
 #' #Renaming (because why not?)
-#' crosstable(mtcars2, by=vs, total="both", test=TRUE, effect=TRUE) %>%
+#' crosstable(mtcars2, -model, by=vs, total="both", test=TRUE, effect=TRUE) %>%
 #'    rename(ID=.id, math=variable, Tot=Total, lab=label, pval=test, fx=effect) %>%
 #'    as_flextable(by_header = "Engine shape", 
 #'                 generic_labels=list(id = "ID", variable = "math", total="Tot", 
@@ -56,31 +57,34 @@ as_flextable.crosstable = function(x, keep_id=FALSE, by_header=NULL,
                                    show_test_name=TRUE, 
                                    fontsizes=list(body=11, subheaders=11, header=11), 
                                    padding_v=NULL, remove_header_keys=FALSE,
-                                   header_show_n=FALSE, 
+                                   header_show_n=FALSE, header_show_n_pattern="{.col} (N={.n})", 
                                    generic_labels=list(id=".id", variable="variable", value="value", 
                                                        total="Total", label="label", test="test", 
                                                        effect="effect"), 
                                    ...) {
     assert_class(x, "crosstable", .var.name=vname(x))
     
-    if(missing(keep_id)) keep_id = getOption("crosstable_keep_id", FALSE)
-    if(missing(autofit)) autofit = getOption('crosstable_autofit', TRUE)
-    if(missing(compact)) compact = getOption('crosstable_compact', FALSE)
-    if(missing(show_test_name)) show_test_name = getOption('crosstable_show_test_name', TRUE)
-    if(missing(padding_v)) padding_v = getOption('crosstable_padding_v', NULL)
-    if(missing(header_show_n)) header_show_n = getOption('crosstable_header_show_n', FALSE)
-    if(missing(remove_header_keys)) remove_header_keys = getOption('crosstable_remove_header_keys', FALSE)
-    # if(missing(generic_labels)) generic_labels = getOption('crosstable_generic_labels', NULL)
+    if(missing(keep_id)) keep_id = getOption("crosstable_keep_id", keep_id)
+    if(missing(autofit)) autofit = getOption('crosstable_autofit', autofit)
+    if(missing(compact)) compact = getOption('crosstable_compact', compact)
+    if(missing(show_test_name)) show_test_name = getOption('crosstable_show_test_name', show_test_name)
+    if(missing(padding_v)) padding_v = getOption('crosstable_padding_v', padding_v)
+    if(missing(remove_header_keys)) remove_header_keys = getOption('crosstable_remove_header_keys',
+                                                                   remove_header_keys)
+    if(missing(header_show_n)) header_show_n = getOption('crosstable_header_show_n', header_show_n)
+    if(missing(header_show_n_pattern)) header_show_n_pattern = getOption('crosstable_header_show_n_pattern', header_show_n_pattern)
+    if(missing(generic_labels)) generic_labels = getOption('crosstable_generic_labels', generic_labels)
     if(missing(fontsizes)) fontsizes = list(
-        body=getOption('crosstable_fontsize_body', 11),
-        subheaders=getOption('crosstable_fontsize_subheaders', 11),
-        header=getOption('crosstable_fontsize_header', 11)
+        body=getOption('crosstable_fontsize_body', fontsizes$body),
+        subheaders=getOption('crosstable_fontsize_subheaders', fontsizes$subheaders),
+        header=getOption('crosstable_fontsize_header', fontsizes$header)
     )
     
     border1 = fp_border(color = "black", style = "solid", width = 1)
     border2 = fp_border(color = "black", style = "solid", width = 1.5)
     labs.names = setdiff(names(x), generic_labels)
     
+    xattr = attributes(x)
     has_test = attr(x, "has_test")
     has_effect = attr(x, "has_effect")
     has_total = attr(x, "has_total")
@@ -92,11 +96,12 @@ as_flextable.crosstable = function(x, keep_id=FALSE, by_header=NULL,
         if(showNA=="always") .x=unique(c(.x, NA))
         replace_na(.x, replace="NA")
     })
+    n_levels = length(by_levels)
     by = attr(x, "by")
     has_by =  !is.null(by)
     if(has_by && is.null(by_label)) by_label=by
-    # if(showNA=="always") by_levels=c(by_levels, "NA")
-    
+    if(isTRUE(by_header)) by_header=NULL
+    if(identical(by_header, "")) by_header=FALSE
     
     test=generic_labels$test
     id=generic_labels$id
@@ -106,24 +111,21 @@ as_flextable.crosstable = function(x, keep_id=FALSE, by_header=NULL,
         x[[test]] = str_remove(x[[test]], "\\n\\(.*\\)")
     }
     if (compact && !inherits(x, "compacted_crosstable")) {
-        x = compact.crosstable(x)
+        x = compact.crosstable(x, keep_id=keep_id)
     }
     
     rtn = replace(x, is.na(x), "NA")
-    if(header_show_n){
-        #TODO faire ça en multiby !
-        col_names_n = imap_chr(c(by_table), ~glue("{.y} (n={.x})")) 
-        # rtn = rtn %>% rename_with(~ifelse(.x %in% names(col_names_n), col_names_n[.x], .x))
-    }
     
     if(inherits(x, "compacted_crosstable")) {
-        # if(length(by_levels)>1) abort("Cannot compact a crosstable with multiple `by`.")
         rows = attr(x, "title_rows")
         title_rows = which(rows)+(0:(sum(rows)-1))
         padded_rows = 1:nrow(rtn)
         padded_rows = padded_rows[!padded_rows %in% title_rows]
+        header_labels = names(rtn) %>% recode(!!generic_labels$variable:="")
+        hl = set_names("", generic_labels$variable)
         rtn = rtn %>% 
             flextable() %>% 
+            set_header_labels(values=hl) %>% 
             fontsize(size=fontsizes$body) %>%
             fontsize(i=title_rows, size=fontsizes$subheaders) %>%
             border(title_rows, border.top = fp_border()) %>%
@@ -134,10 +136,9 @@ as_flextable.crosstable = function(x, keep_id=FALSE, by_header=NULL,
         sep.rows = which(rtn[[id]] != lead(rtn[[id]]))
         body_merge = intersect(names(rtn), generic_labels[c("label", "test", "effect", "id")]) %>%
             unlist()
+        cols = rtn %>% names()
         
-        if(keep_id) {
-            cols = rtn %>% names()
-        } else {
+        if(!keep_id) {
             cols = rtn %>% select(-any_of(id)) %>% names()
             body_merge = body_merge[body_merge!=id]
         }
@@ -152,7 +153,7 @@ as_flextable.crosstable = function(x, keep_id=FALSE, by_header=NULL,
             merge_v(j=id, target=body_merge, part = "body")
     }
     
-    if(length(by_levels)==1) {
+    if(n_levels==1) {
         by_levels2 = unlist(by_levels)
         byname = if(has_label) by_label else by
         
@@ -162,45 +163,68 @@ as_flextable.crosstable = function(x, keep_id=FALSE, by_header=NULL,
             .col_1 = col_keys
         )
         
-        if(!is.null(by_header)){
-            header_mapping = header_mapping %>% 
-                mutate(.col_2=str_replace(.col_2,byname, by_header))
+        if(inherits(x, "compacted_crosstable")) {
+            header_mapping[header_mapping$col_keys==generic_labels$variable, -1] = ""
         }
-        
-        if(header_show_n){
+        if(!isFALSE(header_show_n)){
             header_mapping = header_mapping %>% 
                 mutate(
-                    n = by_table[col_keys], 
-                    .col_1 = ifelse(!is.na(n), glue("{.col_1} (n={n})"), .col_1)
+                    .n = by_table[col_keys], 
+                    .col_1 = ifelse(!is.na(.n), glue(header_show_n_pattern, .col=.col_1), .col_1)
                 ) %>% 
-                select(-n)
+                select(-.n)
         }
-        rtn = rtn %>% 
-            set_header_df(header_mapping, key = "col_keys") %>%
-            merge_h(part = "head")
+        
+        if(!isFALSE(by_header)){
+            if(!is.null(by_header)){
+                header_mapping = header_mapping %>% 
+                    mutate(.col_2=str_replace(.col_2, byname, by_header))
+            }
+            rtn = rtn %>% 
+                set_header_df(header_mapping, key = "col_keys") %>%
+                merge_h(part = "head")
+        }
         
         
-    } else if(length(by_levels)>1) {
+    } else if(n_levels>1) {
+        if(!is.null(by_header)){
+            warn("by_header is ignored if the crosstable has several `by` stratum.", 
+                 class = "crosstable_asflex_byheader_multi")
+        }
         header_mapping = tibble(col_keys = names(x)) %>% 
-            separate(col_keys, into=paste0(".col_", 1:length(by_levels)), 
+            separate(col_keys, into=paste0(".col_", seq(n_levels)), 
                      sep=" & ", remove=FALSE, fill="right") %>% 
-            select(col_keys, rev(names(.))) %>% 
-            mutate(across(starts_with(".col_"), ~ifelse(is.na(.x), col_keys, .x)))
+            mutate(across(starts_with(".col_"), ~ifelse(is.na(.x), col_keys, .x))) %>% 
+            select(col_keys, rev(names(.))) 
         
-        #TODO as_flextable(remove_header_keys) plutôt que logical, pattern avec key, label et value
-        #par exemple pattern_multiby="{.key}={.value}" et pattern_n="" ou "{value} (n={n})"
-        #TODO pattern_n sur toutes les lignes en multiby ?
+        if(!isFALSE(header_show_n)){
+            if(isTRUE(header_show_n)) header_show_n = seq(n_levels)
+            header_show_n = as.numeric(header_show_n)
+            .cols = paste0(".col_", seq(n_levels))
+            header_show_n2 = .cols[as.numeric(header_show_n)]
+            for(i in seq_along(.cols)){
+                .c = .cols[i:length(.cols)]
+                header_mapping = header_mapping %>%
+                    group_by(across(all_of(.c))) %>%
+                    mutate(!!.c[1] := {
+                        .col = !!sym(.c[1])
+                        .col2 = str_split(.col, "=")[[1]]
+                        .col_key = .col2[1]
+                        .col_val = .col2[2]
+                        .n = sum(by_table[col_keys])
+                        ifelse(!is.na(.n) & .c[1] %in% header_show_n2, 
+                               glue(header_show_n_pattern), .col)
+                    })
+            }
+            
+            header_mapping %>% 
+                ungroup() %>% 
+                select(-starts_with("n"))
+        }
+        
         if(remove_header_keys){ 
             header_mapping = header_mapping %>% 
                 mutate(across(starts_with(".col_"), ~str_remove(.x, "^.*=")))
-        }
-        if(header_show_n){
-            header_mapping = header_mapping %>% 
-                mutate(
-                    n = by_table[col_keys], 
-                    .col_1 = ifelse(!is.na(n), glue("{.col_1} (n={n})"), .col_1)
-                ) %>% 
-                select(-n)
         }
         border_left_first = sum(rtn$header$col_keys %in% generic_labels[c("label", "variable", "id")])
         border_separations = header_mapping %>% 
@@ -210,7 +234,7 @@ as_flextable.crosstable = function(x, keep_id=FALSE, by_header=NULL,
         
         rtn = rtn %>% 
             set_header_df(header_mapping, key = "col_keys") %>%
-            merge_h(i=seq.int(length(by_levels)-1), part = "head") %>%
+            merge_h(i=seq.int(n_levels-1), part = "head") %>%
             border(j=borders_j, border.left=border1, part="all") %>%
             vline_left(border=border1) %>% 
             vline_right(border=border1)