+++ R/labels.R
@@ -150,6 +150,7 @@ remove_label = remove_labels
 #' Rename every column of a dataframe with its label
 #'
 #' @param df a data.frame
+#' @param except columns that should not be renamed.
 #' 
 #' @return A dataframe, as `df`, which names are copied from the label attribute
 #'
@@ -158,13 +159,13 @@ remove_label = remove_labels
 #' @export
 #'
 #' @examples
-#' library(dplyr)
-#' mtcars2 %>% 
-#'   select(1:5) %>% 
-#'   rename_dataframe_with_labels()
-rename_dataframe_with_labels = function(df){
+#' rename_dataframe_with_labels(mtcars2[,1:5], except=5) %>% names
+#' rename_dataframe_with_labels(iris2, except=Sepal.Length) %>% names
+rename_dataframe_with_labels = function(df, except=NULL){
     assert_data_frame(df, null.ok=TRUE)
-    names(df) = get_label(df)
+    except = eval_select(enquo(except), data=df)
+    if(length(except)==0) except = ncol(df)+1
+    names(df)[-except] = get_label(df)[-except]
     df
 }