+++ R/officer.R
@@ -58,13 +58,17 @@ body_add_crosstable = function (doc, x, body_fontsize=NULL,
 
 #' Add a new paragraph with default style 
 #' 
-#' Add a new paragraph in an `officer` document with default style. Variables can be inserted as multiple strings (`paste()` style) or enclosed by braces (`glue()` style). 
-#' References to any bookmark can be inserted using the syntax "\\@ref(bookmark)". See an example in [body_add_table_legend()].
+#' Add a new paragraph in an `officer` document with default style.\cr
+#' Variables can be inserted in the text as multiple strings (`paste()` style) or enclosed by braces (`glue()` style).  \cr
+#' Basic markdown syntax is available: `**bold**`, `*italic*`, and `_underlined_`. \cr
+#' References to any bookmark can be inserted using the syntax `\\@ref(bookmark)`.
 #'
 #' @param doc the doc object (created with the `read_docx` function of `officer` package)
 #' @param ... one or several character strings, pasted using `.sep`. As with `glue::glue()`, expressions enclosed by braces will be evaluated as R code. If more than one variable is passed, all should be of length 1.
+#' @param style Style for normal text. Best set with [crosstable_options()].
 #' @param .sep Separator used to separate elements.
 #' @param squish Whether to squish the result (remove trailing and repeated spaces). Default to `TRUE`. Allows to add multiline paragraph without breaking the string.
+#' @param parse which format to parse. Default to all formats (`c("ref", "format", "code")`).
 #'
 #' @return a new doc object
 #' 
@@ -72,7 +76,6 @@ body_add_crosstable = function (doc, x, body_fontsize=NULL,
 #' @export
 #' @importFrom glue glue glue_collapse
 #' @importFrom officer body_add_par
-#' @importFrom purrr map_dbl
 #' @importFrom stringr str_squish
 #' 
 #' @return The docx object `doc`
@@ -86,29 +89,41 @@ body_add_crosstable = function (doc, x, body_fontsize=NULL,
 #' doc = read_docx()  %>% 
 #'     body_add_normal("Table iris has", ncol(iris), "columns.", .sep=" ") %>% #paste style
 #'     body_add_normal("However, table mtcars has {ncol(mtcars)} columns") %>% #glue style
-#'     body_add_normal(info_rows)                                              #vector style
-#' #write_and_open(doc)
-body_add_normal = function(doc, ..., .sep="", squish=TRUE) {
+#'     body_add_normal(info_rows)                                          %>% #vector style
+#'     body_add_normal("")                                              
+#' doc = doc %>% 
+#'     body_add_normal("You can write text in *italic1*, _underlined1_, **bold1**, and `code`, 
+#'                     and you can also add * **references** *, for instance a ref to Table 
+#'                     \\@ref(my_table). Multiple spaces are ignored (squished) so that you 
+#'                     can enter multiline text.") %>% 
+#'     body_add_normal() %>% 
+#'     body_add_normal("Here I should use `body_add_crosstable()` to add a table before the 
+#'                      legend.") %>% 
+#'     body_add_table_legend("My pretty table", bookmark="my_table")
+#' write_and_open(doc)
+body_add_normal = function(doc, ..., .sep="", style=NULL, squish=TRUE, parse=c("ref", "format", "code")) {
     if(missing(squish)) squish = getOption("crosstable_normal_squish", TRUE)
     dots = list(...)
-    normal_style = getOption('crosstable_style_normal', doc$default_styles$paragraph)
-    lengths = map_dbl(dots, length)
+    if(is.null(style)){
+        style = getOption('crosstable_style_normal', doc$default_styles$paragraph)
+    }
+    if(length(dots)==0) dots=""
+    dots_lengths = lengths(dots)
     
-    if(all(lengths==1)){ #one or several vectors of length 1
-        value = glue(..., .sep=.sep, .envir=parent.frame())
+    if(all(dots_lengths==1)){ #one or several vectors of length 1
+        value = do.call(glue, c(dots, .sep=.sep, .envir=parent.frame()))
         if(squish) value = str_squish(value)
-        if(length(value) > 0 && str_detect(value, "\\\\@ref\\((.*?)\\)")){
-            doc = body_add_par(doc, "") %>% parse_reference(value)
-        } else{
-            doc = body_add_par(doc, value, style=normal_style)
-        }
+        parse_ref = "ref" %in% parse
+        parse_format = "format" %in% parse
+        parse_code = "code" %in% parse
+        doc = body_add_parsed(doc, value, style, parse_ref, parse_format, parse_code)
     } else if(length(dots)==1) { #one vector (of 1 or more) -> recursive call
         for(i in dots[[1]]){
             doc = body_add_normal(doc, i, .sep=.sep, squish=squish)
         }
     } else { #several vectors of which at least one is length 2+
         abort(c("body_add_normal() only accepts either one vector of any length or several vectors of length 1", 
-                i=glue("Length of vectors passed: {glue_collapse(lengths, ', ')}")),
+                i=glue("Length of vectors passed: {glue_collapse(dots_lengths, ', ')}")),
               class="officer_wrong_vector_error")
     }
     
@@ -148,7 +163,8 @@ body_add_title = function(doc, value, level = 1, squish=TRUE,
     value = glue(value, .envir = parent.frame())
     if(squish) value = str_squish(value)
     style = paste(style, level)
-    body_add_par(doc, value, style = style)
+    # body_add_par(doc, value, style = style)
+    body_add_parsed(doc, value, style = style)
 }
 
 
@@ -206,7 +222,7 @@ body_add_list_item = function(doc, value, ordered=FALSE, style=NULL, ...){
                   class="officer_lists_style_error") #nocov
         }
     }
-    body_add_par(doc, value, style=style, ...)
+    body_add_parsed(doc, value, style=style, ...)
 }
 
 
@@ -303,6 +319,7 @@ body_add_flextable_list = body_add_crosstable_list
 #' @param doc a docx object
 #' @param legend the table legend. As with [glue::glue()], expressions enclosed by braces will be evaluated as R code.
 #' @param bookmark the id of the bookmark. This is the id that should then be called in [body_add_normal()] using the `"\\@ref(id)"` syntax.
+#' @param legend_prefix a prefix that comes before the legend, after the numbering
 #' @param legend_style style of of the whole legend. May depend on the docx template. However, if `name_format` is provided with a specific `font.size`, this size will apply to the whole legend for consistency.
 #' @param name_format format of the legend's LHS (legend_name + numbering) using [officer::fp_text_lite()] or [officer::fp_text()]. Default to `fp_text_lite(bold=TRUE)` in addition to the format defined in `legend_style`. Note that the reference to the bookmark will have the same specific format in the text.
 #' @param legend_name name before the numbering. Default to either "Table" or "Figure".
@@ -310,6 +327,7 @@ body_add_flextable_list = body_add_crosstable_list
 #' @param seqfield Keep default. Otherwise, you may figure it out doing this: in a docx file, insert a table legend, right click on the inserted number and select "Toggle Field Codes". This argument should be the value of the field, with extra escaping.
 #' @param par_before,par_after should an empty paragraph be inserted before/after the legend?
 #' @param legacy use the old version of this function, if you cannot update `{officer}` to v0.4+
+#' @param ... unused
 #' 
 #' @return The docx object `doc`
 #'
@@ -351,39 +369,45 @@ body_add_flextable_list = body_add_crosstable_list
 #' write_and_open(x)
 #' #If asked to update fields, press "Yes". Otherwise press Ctrl+A then F9 twice for the references 
 #' #to appear.
-body_add_table_legend = function(doc, legend, bookmark=NULL, 
+body_add_table_legend = function(doc, legend, ..., bookmark=NULL, 
                                  legend_style=getOption('crosstable_style_legend', 
                                                         doc$default_styles$paragraph), 
                                  style=deprecated(), 
+                                 legend_prefix=NULL,
                                  name_format=NULL,
                                  legend_name="Table", 
                                  seqfield="SEQ Table \\* Arabic", 
                                  par_before=FALSE,
                                  legacy=FALSE){
+    check_dots_empty()
     if(missing(par_before)) par_before = getOption("crosstable_table_legend_par_before", FALSE)
+    if(missing(legend_prefix)) legend_prefix = getOption("crosstable_table_legend_prefix", NULL)
     if(par_before){
         doc=body_add_normal(doc, "")
     }
     body_add_legend(doc=doc, legend=legend, legend_name=legend_name,
-                    bookmark=bookmark, legend_style=legend_style,
+                    bookmark=bookmark, legend_prefix=legend_prefix, legend_style=legend_style,
                     name_format=name_format, seqfield=seqfield, 
                     style=style, legacy=legacy)
 }
 
 #' @rdname body_add_legend
 #' @export
-body_add_figure_legend = function(doc, legend, bookmark=NULL, 
+body_add_figure_legend = function(doc, legend, ..., bookmark=NULL, 
                                   legend_style=getOption('crosstable_style_legend', 
                                                          doc$default_styles$paragraph), 
                                   style=deprecated(), 
+                                  legend_prefix=NULL,
                                   name_format=NULL,
                                   legend_name="Figure", 
                                   seqfield="SEQ Figure \\* Arabic", 
                                   par_after=FALSE,
                                   legacy=FALSE){
+    check_dots_empty()
     if(missing(par_after)) par_after = getOption("crosstable_figure_legend_par_after", FALSE)
+    if(missing(legend_prefix)) legend_prefix = getOption("crosstable_figure_legend_prefix", NULL)
     doc = body_add_legend(doc=doc, legend=legend, legend_name=legend_name,
-                    bookmark=bookmark, legend_style=legend_style,
+                    bookmark=bookmark, legend_prefix=legend_prefix, legend_style=legend_style,
                     name_format=name_format, seqfield=seqfield, 
                     style=style, legacy=legacy)
     if(par_after){
@@ -398,7 +422,7 @@ body_add_figure_legend = function(doc, legend, bookmark=NULL,
 #' @keywords internal
 #' @noRd
 body_add_legend = function(doc, legend, legend_name, bookmark, 
-                           legend_style, name_format, seqfield, 
+                           legend_prefix, legend_style, name_format, seqfield, 
                            style, legacy){
     
     # nocov start
@@ -412,6 +436,7 @@ body_add_legend = function(doc, legend, legend_name, bookmark,
             style = getOption('crosstable_style_strong', "strong")
         }
         
+        
         rtn = body_add_legend_legacy(doc=doc, legend=legend, legend_name=legend_name,
                                bookmark=bookmark, legend_style=legend_style, style=style, seqfield=seqfield)
         return(rtn)
@@ -420,13 +445,14 @@ body_add_legend = function(doc, legend, legend_name, bookmark,
     if(is_present(style)){
         deprecate_warn("0.2.2", "body_add_X_legend(style)", 
                        "body_add_X_legend(name_format)", 
-                       details="The `style` argument has been ignored. Use `legacy=TRUE` to override.")
+                       details="Therefore, its value has been ignored. Use `legacy=TRUE` to override.")
     }
     # nocov end
-
+    
+    legend = paste0(legend_prefix, legend)
     fp_text2 = officer::fp_text_lite #v0.4+
     if(is.null(name_format)){
-        name_format = fp_text2(bold=TRUE)
+        name_format = getOption('crosstable_format_legend_name', fp_text2(bold=TRUE))
     }
     fp_size = fp_text2(font.size=name_format$font.size)
     
@@ -456,7 +482,6 @@ body_add_legend = function(doc, legend, legend_name, bookmark,
 #' @noRd
 body_add_legend_legacy = function(doc, legend, legend_name, bookmark, 
                                   legend_style, style, seqfield){
-    # browser()
     legend = glue(legend, .envir = parent.frame())
     rtn = doc %>% 
         body_add_par(value=legend, style=legend_style) %>% 
@@ -740,95 +765,80 @@ generate_autofit_macro = function(){
 # Internal utils ---------------------------------------------------------
 
 
-#' Recursive helper function 
-#' Replace every string containing a reference to a table/figure by the 
-#' docx-formatted cross-reference
+#' Parse value for multiple regexp to unravel formats (bold, italic and underline) and reference calls.
 #' 
 #' @importFrom stringr str_split str_detect str_match str_extract_all
 #' @importFrom glue glue
 #' @importFrom utils packageVersion
-#' @importFrom purrr map
-#' @importFrom officer run_word_field ftext body_add_fpar
+#' @importFrom purrr map map_lgl discard set_names
+#' @importFrom rlang warn
+#' @importFrom officer run_word_field ftext body_add_fpar fp_text_lite
 #' 
 #' @keywords internal
 #' @noRd
-parse_reference = function(doc, value){
+body_add_parsed = function(doc, value, style, parse_ref=TRUE, parse_format=TRUE, parse_code=TRUE){
     if(packageVersion("officer")<"0.4"){
-        return(parse_reference_legacy(doc, value))
-    }
-    par_not_ref = str_split(value, "\\\\@ref\\(.*?\\)")[[1]]
-    par_ref = stringr::str_extract_all(value, "\\\\@ref\\(.*?\\)")[[1]]
-    #altern: https://stackoverflow.com/a/43876294/3888000
-    altern = c(par_not_ref, par_ref)[order(c(seq_along(par_not_ref)*2 - 1, seq_along(par_ref)*2))] 
-    
-    par_list = map(altern, ~{
-        if(str_detect(.x, "\\\\@ref")){
-            bkm = stringr::str_match(.x, "\\\\@ref\\((.*?)\\)")[,2]
-            run_word_field(glue(' REF {bkm} \\h '))
-        } else {
-            ftext(.x)
-        }
-    })
-    
-    p=do.call(fpar, args=par_list)
-    body_add_fpar(doc, p)
-}
-
-
-#' TODO: faire une fonction qui parse le format (**=gras, *=italique, _=souligné)
-#' @keywords internal
-#' @noRd
-parse_format = function(doc, value){
-    if(packageVersion("officer")<"0.4") return(doc)
-    browser()
-    par_not_ref = str_split(value, "\\\\@ref\\(.*?\\)")[[1]]
-    par_ref = stringr::str_extract_all(value, "\\\\@ref\\(.*?\\)")[[1]]
-    #altern: https://stackoverflow.com/a/43876294/3888000
-    altern = c(par_not_ref, par_ref)[order(c(seq_along(par_not_ref)*2 - 1, seq_along(par_ref)*2))] 
-    
-    par_list = map(altern, ~{
-        if(str_detect(.x, "\\\\@ref")){
-            bkm = stringr::str_match(.x, "\\\\@ref\\((.*?)\\)")[,2]
-            run_word_field(glue(' REF {bkm} \\h '))
-        } else {
-            ftext(.x)
-        }
-    })
-    
-    p=do.call(fpar, args=par_list)
-    body_add_fpar(doc, p)
-}
-
-# nocov start
-
-#' @importFrom stringr str_detect str_match_all
-#' @importFrom glue glue
-#' @importFrom rlang abort
-#' @importFrom officer slip_in_text slip_in_seqfield
-#' @keywords internal
-#' @noRd
-parse_reference_legacy = function(doc, value){
-    normal_style_character = getOption('crosstable_style_character', doc$default_styles$character)
-    
-    if(!str_detect(value, "\\\\@ref\\((.*?)\\)")){ #recursion out
-        doc = doc %>% 
-            slip_in_text(value, style=normal_style_character, pos='after')
+        warn("This function needs package {officer} v0.4+ to work. You won't be able to add formatted text or references until you update this package.")
         return(doc)
     }
+    if(isFALSE(parse_ref) && isFALSE(parse_format)){
+        body_add_par(doc, value, style)
+    }
+    reg_r = list(
+        ref = "\\\\@ref\\(.*?\\)"
+    )
+    reg_f = list(
+        bold = "\\*\\*(.+?)\\*\\*",
+        underlined = "_(.+?)_",
+        italic = "(?<!\\*)\\*(?!\\*)(.+?)(?<!\\*)\\*(?!\\*)"
+    )
+    reg_c = list(
+        code = "`(.+?)`"
+    )
+    if(isFALSE(parse_ref)) reg_r = list()
+    if(isFALSE(parse_format)) reg_f = list()
+    if(isFALSE(parse_code)) reg_c = list()
+    regex = c(reg_f, reg_r, reg_c)
+    rex_all = paste(regex, collapse="|")
     
-    x = str_match_all(value, "(.*?)\\\\@ref\\((.*?)\\)(.*)")
-    x = unlist(x[[1]][-1])
-    if(length(x)!=3) #1rd (.*) is not greedy
-        abort(c("This error should not have happened, please report a bug.", # nocov
-                i="function = crosstable:::parse_reference"), x=x)           # nocov
-    doc = doc %>% 
-        slip_in_text(x[1], style=normal_style_character, pos='after') %>% 
-        slip_in_seqfield(str = glue(' REF {x[2]} \\h '),  
-                         style=normal_style_character, pos='after') %>%
-        parse_reference(x[3])
+    par_not_format = str_split(value, rex_all)[[1]]
+    par_format = str_extract_all(value, rex_all)[[1]]
+    
+    # #altern: https://stackoverflow.com/a/43876294/3888000
+    altern = c(par_not_format, par_format)[order(c(seq_along(par_not_format)*2 - 1, 
+                                                   seq_along(par_format)*2))]
+    par_list = map(altern, ~{
+        .format = map_lgl(regex, function(pat) str_detect(.x, pattern=pat)) %>% 
+            discard(isFALSE) %>% names()
+        
+        if(length(.format)==0) return(ftext(.x))
+        
+        if(any(.format=="ref")){
+            bkm = str_match(.x, "\\\\@ref\\((.*?)\\)")[,2]
+            return(run_word_field(glue(' REF {bkm} \\h ')))
+        }
+        if(any(.format=="code")){
+            fp = fp_text_lite(font.family=getOption("crosstable_font_code", "Consolas"))
+            return( ftext(.x, fp))
+        }
+        rex = regex[.format]
+        for(i in rex){
+            if(str_detect(.x, i)){
+                .x = str_match(.x, i)[[2]]
+            }
+        }
+        
+        fp_args = rep(TRUE, length(.format)) %>% set_names(.format) %>% as.list()
+        fp = do.call(fp_text_lite, fp_args)
+        
+        ftext(.x, fp)
+    })
+    
+    
+    p=do.call(fpar, args=par_list)
+    body_add_fpar(doc, p, style)
 }
 
-# nocov end
 
 
 
@@ -838,7 +848,6 @@ parse_reference_legacy = function(doc, value){
 #' @usage NULL
 #' @importFrom lifecycle deprecate_warn
 #' @rdname body_add_normal
-#' @author Dan Chaltiel
 #' @export
 body_add_glued = function(...){
     deprecate_warn("0.2.0", "body_add_glued()", "body_add_normal()")# nocov