+++ R/options.R
@@ -13,6 +13,12 @@
 #' @param crosstable_units default units in [body_add_gg2()] and [body_add_img2()]
 #' @param crosstable_peek_docx behavior of [peek()], which will open a `docx` if `TRUE` (default) and an `xlsx` if `FALSE`
 #' @param crosstable_compact_padding in flextables, left-padding for non-headers rows when `compact=TRUE`.
+#' @param crosstable_font_code font family used to show code, most likely a monospaced typeface such as Consolas (default)
+#' @param crosstable_format_legend_name how the legend name ("Table", "Figure") is formatted. Default to `officer::fp_text_lite(bold=TRUE)`
+#' @param crosstable_table_legend_par_before whether to add an empty paragraph before all table legends
+#' @param crosstable_figure_legend_par_after whether to add an empty paragraph after all figure legends
+#' @param crosstable_table_legend_prefix,crosstable_figure_legend_prefix a prefix before each legend, after the numbering
+#' @param crosstable_header_show_n_pattern glue pattern used when showing N in the header of flextables. `.col` is the name of the column and `.n` the size of the group. Default to `{.col} (N={.n})`.
 #' 
 #' @param crosstable_total For setting [crosstable()] arguments globally. 
 #' @param crosstable_percent_pattern For setting [crosstable()] arguments globally. 
@@ -62,6 +68,13 @@ crosstable_options = function(reset=FALSE,
                               crosstable_units, 
                               crosstable_peek_docx,
                               crosstable_compact_padding, 
+                              crosstable_font_code, 
+                              crosstable_format_legend_name,
+                              crosstable_table_legend_par_before, 
+                              crosstable_table_legend_prefix, 
+                              crosstable_figure_legend_par_after, 
+                              crosstable_figure_legend_prefix,
+                              crosstable_header_show_n_pattern, 
                               #crosstable()
                               crosstable_total, crosstable_percent_pattern, crosstable_percent_digits, crosstable_num_digits, crosstable_showNA, crosstable_label, crosstable_funs, crosstable_funs_arg, crosstable_cor_method, crosstable_unique_numeric, crosstable_date_format, crosstable_times, crosstable_followup, crosstable_test_arg, crosstable_effect_args, 
                               #as_flextable()